move template handling to utility package so can be reused for SMS templates as well
